ads1292 test # of output are lacked

+++ khu_sensor/Source/Filter_Test/Filter_Test_Verilog/ADS1292_Filter/ADS1292_Filter_CPP/ads1292_filter.cpp
@@ -100,7 +100,7 @@ int ads1292_filter::write_Passed_Data(const std::string &t_FileName){
             return -1;
         } else {
             if(this->m_Signal != nullptr) {
-                float temp = 0.0; // signal value
+                signed int temp = 0.0; // signal value
                 unsigned int code = 0; // for hex code
                 DoublyIterator<signed int> s_iter(*this->m_Signal);
                 while (!s_iter.NextIsTail()){
